fix issue with permalinks of retweeted tweets
basically, retweets have a retweet id and an original id. they need to have the retweet id as their real
ID, because otherwise the ID-based session stuff gets funky.

but since that means the post author and the post ID won't match, need to generate the permalink using the
screen name of the retweeting user rather than the original user. twitter.com will then redirect to the original
tweet and display the retweeted status.